requirements.txt
pytz==2022.7.1
requests==2.28.2
python-dateutil==2.8.2
python-telegram-bot==20.1
